requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
email-validator>=2.0.0
stripe>=8.0.0
svix>=1.80.0
